Add `RevitRuleExtensionTest` test for `FilterRuleExtension`.
